Добавлены MetricNumber, MetricNumberImage, EJB и тест для них. А также начата реализация Product.

+++ src/main/java/ru/kiasystems/logic/beans/impl/EmployeeEJB.java
@@ -35,7 +35,7 @@ public class EmployeeEJB implements EmployeeEJBRemote {
         entityManager.persist(theme);
     }
     @Override
-    public Employee getEmployeeById(int id) {
+    public Employee getEmployeeById(Long id) {
         return entityManager.find(Employee.class, id);
     }
     @Override
