+++ src/main/java/ru/kiasystems/model/entity/User.java
@@ -11,7 +11,8 @@ public class User implements Serializable{
 
     @Id
     @Column(name="employee_id")
-    private Integer id;
+    private Long id;
+
     @Column(name="username", length = 20, nullable = false)
     private String username;
 
@@ -26,11 +27,11 @@ public class User implements Serializable{
 		
     public User(){}
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
